requirements.txt
alembic==0.9.8
celery==4.1.0
coverage==4.5.1
Flask==0.12.2
Flask-JWT-Extended==3.7.2
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
html2text==2018.1.9
itsdangerous==0.24
Jinja2==2.10
mistune==0.8.3
Mako==1.0.7
MarkupSafe==1.0
passlib==1.7.1
patool==1.12
psycopg2-binary==2.7.4
py==1.5.2
pytest==3.4.2
pytest-cov==2.5.1
python-dateutil==2.7.0
python-editor==1.0.3
six==1.11.0
SQLAlchemy==1.2.5
SQLAlchemy-Utils==0.33.0
Werkzeug==0.13
python-archive==0.2
flake8==3.5.0
pylint==1.8.2
git+https://github.com/i-kiwamu/python3-oauth2.git
click==6.7
lxml==4.1.1
pyflakes==1.6.0
sphinx-rtd-theme==0.2.4
Sphinx==1.7.1
git+https://github.com/libre-man/HTTPdomain#subdirectory=httpdomain
git+https://github.com/libre-man/sphinx-autodoc-typehints
mypy==0.570
mypy-extensions==0.3.0
yapf==0.20.2
isort==4.3.4
PyJWT==1.6.0
validate-email==1.3
typing-extensions==3.6.2.1
Flask-Limiter==1.0.1
